services/auth_service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
uvicorn[standard]